// packages/aa-db/src/Api/Data/Profile.tsx
export interface ProfileComment {
    id: number;
    priority: number;
    condMessage: string;
    comment: string;
    condType: CommentConditionType;
    condValues: number[];
    condValue2: number;
}